backend/auth_service/requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
python-jose
cryptography
pydantic
pydantic[email]
email-validator
pyotp
qrcode
pillow
sqlalchemy
psycopg2-binary
python-multipart
argon2_cffi